[ upd ] switch for telemetry and process pinning

+++ dashboard/dashboard.py
@@ -122,7 +122,7 @@ def target_precheck(root_dir, configs_dir, target_name,
     return ({'success': True, 'message': ''}, target_info)
 
 
-def experiment_precheck(info, experiments_dir, exp_name, default_telemetry_rate):
+def experiment_precheck(info, experiments_dir, exp_name, default_telemetry_rate, top_level_telemetry_switch):
     return target_precheck(
         experiments_dir, info.exp_configs, exp_name,
         {
@@ -131,7 +131,12 @@ def experiment_precheck(info, experiments_dir, exp_name, default_telemetry_rate)
             'rerun_setup': False,
             'tvm_remote': 'origin',
             'tvm_branch': 'master',
-            'telemetry_rate': default_telemetry_rate
+            'telemetry_rate': default_telemetry_rate,
+            'run_telemetry': top_level_telemetry_switch,
+            'process_pinning': {
+                'enable': False,
+                'cores': None
+            }
         },
         ['run.sh', 'analyze.sh', 'visualize.sh', 'summarize.sh'])
 
@@ -200,7 +205,7 @@ def copy_setup(experiments_dir, setup_dir, exp_name):
                     cwd=exp_dir)
 
 
-def run_experiment(info, experiments_dir, tmp_data_dir, exp_name):
+def run_experiment(info, experiments_dir, tmp_data_dir, exp_name, pin_process=False, cores=None, run_telemetry=False):
 
     to_local_time = lambda sec: time.asctime(time.localtime(sec))
     exp_dir = os.path.join(experiments_dir, exp_name)
@@ -215,7 +220,11 @@ def run_experiment(info, experiments_dir, tmp_data_dir, exp_name):
     start_msg = f'Experiment {exp_name} starts @ {to_local_time(start_time)}'
     print_log(start_msg)
     # run the run.sh file on the configs directory and the destination directory
-    subprocess.call([os.path.join(exp_dir, 'run.sh'), exp_conf, exp_data_dir],
+    if pin_process and cores:
+        subprocess.call(['taskset', '--cpu-list', f'{cores}', os.path.join(exp_dir, 'run.sh'), exp_conf, exp_data_dir],
+                    cwd=exp_dir)
+    else:
+        subprocess.call([os.path.join(exp_dir, 'run.sh'), exp_conf, exp_data_dir],
                     cwd=exp_dir)
     end_time = time.time()
     delta = datetime.timedelta(seconds=end_time - start_time)
@@ -231,6 +240,7 @@ def run_experiment(info, experiments_dir, tmp_data_dir, exp_name):
     status['start_time'] = to_local_time(start_time)
     status['end_time'] = to_local_time(end_time)
     status['time_delta'] = str(delta)
+    status['run_telemetry'] = run_telemetry
     # not literally copying because validate may have produced a status that generated an error
     info.report_exp_status(exp_name, 'run', status)
     return status['success']
@@ -341,7 +351,7 @@ def summarize_experiment(info, experiments_dir, exp_name):
 
 def run_all_experiments(info, experiments_dir, setup_dir,
                         tmp_data_dir, data_archive,
-                        time_str, telemetry_script_dir, telemetry_interval=15, randomize=True):
+                        time_str, telemetry_script_dir, run_telemetry=False, telemetry_interval=15, randomize=True):
     """
     Handles logic for setting up and running all experiments.
     """
@@ -354,7 +364,7 @@ def run_all_experiments(info, experiments_dir, setup_dir,
     # do the walk of experiment configs, take account of which experiments are
     # either inactive or invalid
     for exp_name in info.all_present_experiments():
-        precheck, exp_info = experiment_precheck(info, experiments_dir, exp_name, telemetry_interval)
+        precheck, exp_info = experiment_precheck(info, experiments_dir, exp_name, telemetry_interval, run_telemetry)
         info.report_exp_status(exp_name, 'precheck', precheck)
         exp_status[exp_name] = 'active'
         exp_confs[exp_name] = exp_info
@@ -402,14 +412,21 @@ def run_all_experiments(info, experiments_dir, setup_dir,
             tvm_hash = get_tvm_hash()
 
         tvm_hashes[exp] = tvm_hash
-        telemetry_process = start_telemetry(telemetry_script_dir, exp,
-                                            tmp_data_dir,
-                                            interval=exp_confs[exp]['telemetry_rate']
-                                                     if 'telemetry_rate' in exp_confs[exp] 
-                                                     else telemetry_interval)
-        success = run_experiment(info, experiments_dir, tmp_data_dir, exp)
+        pin_process = exp_confs[exp].get('process_pinning', None)
+        print(exp_confs[exp])
+        if run_telemetry and 'run_telemetry' in exp_confs[exp].keys():
+            run_telemetry = run_telemetry and exp_confs[exp]['run_telemetry']
+        enabled = pin_process.get('enable', False) if pin_process else False
+        cores = pin_process.get('cores', None) if enabled else None
+        telemetry_interval = exp_confs[exp].get('telemetry_rate', telemetry_interval)
+        if run_telemetry:
+            telemetry_process = start_telemetry(telemetry_script_dir, exp,
+                                                tmp_data_dir,
+                                                interval=telemetry_interval) if run_telemetry else None
+        success = run_experiment(info, experiments_dir, tmp_data_dir, exp,
+                                 pin_process=pin_process, cores=cores, run_telemetry=run_telemetry)
         # Telemetry can be disabled
-        if telemetry_process:
+        if run_telemetry and telemetry_process:
             telemetry_process.kill()
             # Gather stat collected by the telemetry process
             process_telemetry_statistics(info, exp, tmp_data_dir, time_str)
@@ -548,9 +565,10 @@ def main(home_dir, experiments_dir, subsystem_dir, telemetry_script_dir):
         randomize_exps = dash_config['randomize']
 
     telemetry_rate = dash_config.get('telemetry_rate', 15)
+    run_telemetry = dash_config.get('run_telemetry', False)
     run_all_experiments(info, experiments_dir, setup_dir,
                         tmp_data_dir, data_archive,
-                        time_str, telemetry_script_dir, telemetry_interval=telemetry_rate, randomize=randomize_exps)
+                        time_str, telemetry_script_dir, run_telemetry=run_telemetry, telemetry_interval=telemetry_rate, randomize=randomize_exps)
 
     run_all_subsystems(info, subsystem_dir, time_str)
